feat: generate CLI docs without needing compiled binaries
