+++ cli/cmd.go
@@ -57,10 +57,10 @@ var GetStorageMinerAPI = cliutil.GetStorageMinerAPI
 var GetWorkerAPI = cliutil.GetWorkerAPI
 
 var CommonCommands = []*cli.Command{
-	AuthCmd,
-	LogCmd,
-	WaitApiCmd,
-	FetchParamCmd,
+	WithCategory("developer", AuthCmd),
+	WithCategory("developer", LogCmd),
+	WithCategory("developer", WaitApiCmd),
+	WithCategory("developer", FetchParamCmd),
 	PprofCmd,
 	VersionCmd,
 }
