+++ cli/lotus/backup.go
@@ -1,4 +1,4 @@
-package main
+package lotus
 
 import (
 	"os"
