+++ cli/miner/actor_test.go
@@ -1,4 +1,4 @@
-package main
+package miner
 
 import (
 	"bytes"
