+++ cli/miner/allinfo_test.go
@@ -1,4 +1,4 @@
-package main
+package miner
 
 import (
 	"flag"
