+++ cli/miner/info.go
@@ -1,4 +1,4 @@
-package main
+package miner
 
 import (
 	"context"
