+++ cli/miner/proving.go
@@ -1,4 +1,4 @@
-package main
+package miner
 
 import (
 	"encoding/json"
