+++ cli/miner/sectors.go
@@ -1,4 +1,4 @@
-package main
+package miner
 
 import (
 	"bufio"
