+++ cli/miner/stop.go
@@ -1,4 +1,4 @@
-package main
+package miner
 
 import (
 	_ "net/http/pprof"
