+++ scripts/docsgen-cli/main.go
@@ -3,11 +3,12 @@ package main
 import (
 	"fmt"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
 
-	"golang.org/x/sync/errgroup"
+	"github.com/urfave/cli/v2"
+
+	"github.com/filecoin-project/lotus/cli/lotus"
+	"github.com/filecoin-project/lotus/cli/miner"
+	"github.com/filecoin-project/lotus/cli/worker"
 )
 
 const (
@@ -45,89 +46,25 @@ func main() {
 	}
 
 	fmt.Println("Generating CLI documentation...")
-	var eg errgroup.Group
-	for _, cmd := range []string{"lotus", "lotus-miner", "lotus-worker"} {
-		eg.Go(func() error {
-			err := generateMarkdownForCLI(cmd)
-			if err != nil {
-				fmt.Printf(" ❌ %s: %v\n", cmd, err)
-			} else {
-				fmt.Printf(" ✅ %s\n", cmd)
-			}
-			return err
-		})
+
+	cliApps := map[string]*cli.App{
+		"lotus":        lotus.App(),
+		"lotus-worker": worker.App(),
+		"lotus-miner":  miner.App(),
 	}
-	if err := eg.Wait(); err != nil {
-		fmt.Printf("Failed to generate CLI documentation: %v\n", err)
-		os.Exit(1)
+
+	for name, app := range cliApps {
+		for _, cmd := range app.Commands {
+			cmd.HelpName = fmt.Sprintf("%s %s", app.HelpName, cmd.Name)
+		}
+
+		generator := NewDocGenerator(outputDir, app)
+		if err := generator.Generate(name); err != nil {
+			fmt.Printf(" ❌ %s: %v\n", name, err)
+			continue
+		}
+		fmt.Printf(" ✅ %s\n", name)
 	}
+
 	fmt.Println("Documentation generation complete.")
-}
-
-func generateMarkdownForCLI(cli string) error {
-	md := filepath.Join(outputDir, fmt.Sprintf("cli-%s.md", cli))
-	out, err := os.Create(md)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = out.Close() }()
-	return writeCommandDocs(out, cli, 0)
-}
-
-func writeCommandDocs(file *os.File, command string, depth int) error {
-	// For sanity, fail fast if depth exceeds some arbitrarily large number. In which
-	// case, chances are there is a bug in this script.
-	if depth > depthRecursionLimit {
-		return fmt.Errorf("recursion exceeded limit of %d", depthRecursionLimit)
-	}
-
-	// Get usage from the command.
-	usage, err := exec.Command("sh", "-c", "./"+command+" -h").Output()
-	if err != nil {
-		return fmt.Errorf("failed to run '%s': %v", command, err)
-	}
-
-	// Skip the first new line since the docs do not start with a newline at the very
-	// top.
-	if depth != 0 {
-		if _, err := file.WriteString("\n"); err != nil {
-			return err
-		}
-	}
-
-	// Write out command header and usage.
-	header := fmt.Sprintf("%s# %s\n", strings.Repeat("#", depth), command)
-	if _, err := file.WriteString(header); err != nil {
-		return err
-	} else if _, err := file.WriteString("```\n"); err != nil {
-		return err
-	} else if _, err := file.Write(usage); err != nil {
-		return err
-	} else if _, err := file.WriteString("```\n"); err != nil {
-		return err
-	}
-
-	// Recurse sub-commands.
-	commands := false
-	lines := strings.Split(string(usage), "\n")
-	for _, line := range lines {
-		switch line = strings.TrimSpace(line); {
-		case line == "":
-			commands = false
-		case line == "COMMANDS:":
-			commands = true
-		case strings.HasPrefix(line, "help, h"):
-			// Skip usage command.
-		case commands:
-			// Find the sub command and trim any potential comma in case of alias.
-			subCommand := strings.TrimSuffix(strings.Fields(line)[0], ",")
-			// Skip sections in usage that have no command.
-			if !strings.Contains(subCommand, ":") {
-				if err := writeCommandDocs(file, command+" "+subCommand, depth+1); err != nil {
-					return err
-				}
-			}
-		}
-	}
-	return nil
 }
